Added missing test runner error.
Verify that the test runner exists before attempting to execute it. If
it does not exist, output an error and inform users that they must
perform both a build and install (the default behavior). In other words,
the following command will result in an error:

./build-linux.sh --clean --no-install --test

Additionally, the "-t --test" option has been added the Linux build
script help message.